requirements.txt
Flask==3.0.3
ultralytics==8.3.108
Pillow==10.4.0
supervision==0.25.1
numpy==1.26.1
